Move Kobold API endpoints to their own module
